pyproject.toml
[tool.poetry]
name = "gator-computational-cloud"
version = "0.1.0"
description = ""
authors = ["Nathan Loria <lorian@allegheny.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
python-dotenv = "^0.20.0"
django-widget-tweaks = "^1.4.12"
django-cryptography = "^1.1"
django-recaptcha = "^3.0.0"
mysqlclient = "^2.1.0"
boto3 = "^1.21.35"
dropbox = "^11.29.0"
xmltodict = "^0.12.0"
psutil = "^5.9.0"
paramiko = "^2.10.3"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.1"
flake8 = "^4.0.1"
pylint = "^2.13.4"
taskipy = "^1.10.1"
pytest-codecov = "^0.4.0"
coverage = "^6.3.2"

[tool.taskipy.tasks]
lint = { cmd = "black gcc_exec --check;flake8 gcc_exec --ignore=E501,W503;pylint gcc_exec", help = "Run the black checks for source code format" }
test = { cmd = "pytest --cov gcc_exec/tests --cov-report xml:coverage.xml", help = "Run the test suite to ensure code correctness" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
